% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankStR.R
\name{clean_bank_df}
\alias{clean_bank_df}
\title{Title}
\usage{
clean_bank_df(df, variables, dateformat)
}
\arguments{
\item{df}{The dataframe to clean after importing/reading from a bank's CSV
file.}

\item{variables}{}
}
\value{
A dataframe reduced to the variables "day", "purpose" and "amount" of
the transactions
}
\description{
Title
}

