% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS.R
\name{read_gls}
\alias{read_gls}
\title{Read & clean CSVs from the GLS Gemeinschaftsbank eG}
\usage{
read_gls(file)
}
\arguments{
\item{file}{A CSV file exported from your GLS online account | Finanzen & Ums<U+00E4>tze | Umsatzexport}
}
\value{
A dataframe
}
\description{
Read & clean CSVs from the GLS Gemeinschaftsbank eG
}
\examples{
read_gls('data/gls_Umsaetze_DDMMYYYY_bis_DDMMYYYY_KTO1234567890_DDMMYYYY.csv')
}

