% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparkasse.R
\name{read_sparkasse}
\alias{read_sparkasse}
\title{Read & clean Sparkasse CSVs}
\usage{
read_sparkasse(file)
}
\arguments{
\item{file}{The CSV file that you downlaoded from your Sparkasse under
Postfach | Funktionen | Giro... | Dokumentenuebersicht aufrufen |
Funktionen | Export | Format: CSV...}
}
\value{
A dataframe
}
\description{
Uses latin1 encoding and adjusts column names.
}
\examples{
read_sparkasse('data/19730303-12345678-umsatz.CSV')
}

