% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wustenrot.R
\name{read_wustenrot}
\alias{read_wustenrot}
\title{Read & clean W<U+00FC>stenrot CSVs}
\usage{
read_wustenrot(file)
}
\arguments{
\item{file}{A CSV file exported from your W<U+00FC>stenrot Giro-Konto | Kontoums<U+00E4>tze}
}
\value{
A dataframe
}
\description{
Read & clean W<U+00FC>stenrot CSVs
}
\examples{
read_sparda("data/1234567890_19730303_19730323.csv")
}

