% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparda.R
\name{read_sparda}
\alias{read_sparda}
\title{Read & clean Sparda CSVs}
\usage{
read_sparda(file)
}
\arguments{
\item{file}{A CSV file exported from your Sparda Konto-Center |
Umsaetze | Umsatzsuche}
}
\value{
A dataframe with three columns: "DATE", "DESCRIPTION" and "AMOUNT".
}
\description{
Ignores 10 rows of general account info before table of transactions,
adjusts column names and deletes last row and column which contain only
}
\examples{
read_sparda('data/sparda_umsaetze-1234567-YYYY-MM-DD-HH-MM-SS')

}
